Created an `AnalyticsDelegate` class to support optional delegates
